Fix a bug where the header contact links were un-clickable
If one does not specify a `info->brief` statement (i.e. the first
paragraph of the resume), it seems as if the header for any other
section will overlay the contact section, thus making any of those links
un-clickable.

This works _for me_ because the employment section is the first section
at the top.